Correct webdriver executable fixture in README
Took me a moment to discover that there's no such fixture as `splinter_driver_executable`, so here's a fix. I also clarified that it should return a *directory*, not a path to executable. Although if that is the case, perhaps we should add a `_dir` or `_directory` suffix for fixture name (and CLI option)?